seekhowithrua-lms/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.4
gunicorn==21.2.0
